refactor Movie

+++ app/src/main/java/com/mango/movies/model/domain/Movie.kt
@@ -33,7 +33,7 @@ data class Movie(
     @SerializedName("video")
     val video: Boolean?,
     @SerializedName("vote_average")
-    val voteAverage: Double?,
+    val voteAverage: Float?,
     @SerializedName("vote_count")
     val voteCount: Int?
 ): Parcelable
